[LIGO-232] Test for watched files
Problem: We should ensure that the extension works as intended with
watched files.

Solution: Make a test that creates, changes, and deletes a file, and
checks that all references are correct. Note that I make these three
tests separately as it didn't work in only one for some reason, even
though it still worked as intended in the extension.

On the way, do some housekeeping.