[LIGO-260] Fix trees failing to build
Problem: We are getting frequent "subtrees must be less of [sic] equal"
errors, specially in large codebases such as `ctez`, `baseDAO`,
`segmented-cfmm` and `minter-sdk`.

Solution: Turns out the problem was in the calculation of the
`MarkerInfo`'s location when we have an inclusion. We forgot to subtract
`lmLine` from the previous loc, causing such errors. Everything else in
this MR are just some housekeeping (changing all these `rStart . _1`
with `startLine` for example).