[LIGO-266] Add some parser tests
Problem: It would be nice to test that our changes do work.

Solution: Add some tests, which check that trailing separators in lists
work, and also wildcards. Note that ReasonLIGO seems to have some bug,
and even though it says lists are supported in patterns, it's not
accepting the list.