[LIGO-273] make ligo parsers recoverable
Problem: After extending API of ParserLib we need adjust instantiation of it.
Also create a minimal grammar annotation is required to make error recovery available.

Solution:
- add `RecoverParser` into `PARSER` signature that was extended for error recovery.
- add `'a terminal` type for it into lexers.
- adjust dune configuration to generate recovery information by
`ocaml-recovery-parser` package.
- add dummy recovery annotation to grammars

[LIGO-327] Remove duplication of `'a terminal` type definition

Problem: Since we don't use menhir token and instead of it pass our own
module by `--external-token` we also have to define `'a terminal` indexed type
to satisfy Menhir's Inspection API (that needed for error recovery).
The straightforward solution is just to define it by hand, but it very much resembles
`token` type that we already have.

Solution: use the menhir's option `--only-tokens` to generate a definition of `'a
terminal` and include it into `Token` module in the lexing pass.

Obstacles:
1. menhir generates a module without opening any modules (e. g.
`Simple_utils.Region`). So we need to wrap it to a library with flags `-open` and
use imported symbols as unqualified.

[LIGO-327] Reduce duplication of signatures

Solution:
1. Use MenhirLib.IncrementalEngine.EVERYTHING. It helps to avoid explicitly
connecting INCREMENTAL_ENGINE and INSPECTION (that was added recently).
2. Import the Recovery signature instead of copying it. It requires changes in
the ocaml-recovery-parser that initially doesn't export the signature.