[LIGO-271] Highlight tuple arguments in signature help
Problem: We only highlight arguments `Apply`, but not on tuple
arguments. Since some languages such as CameLIGO heavily rely on tuples
for such a thing, we should also support highlighting them.

Solution: On signature help, we check the case when we only have one
argument in a function, and such argument is a tuple. In this case, we
wish to consider the tuple elements as arguments.

We then need to also check our parameters for `ParameterInformation`.
This turned out to be a tricky case for two reasons. First, PascaLIGO is
a bit divergent on this matter, and it has true function arguments,
unlike ReasonLIGO and CameLIGO which pattern match on arguments. Second,
our `ScopedDecl` parser doesn't really parse patterns, so some work had
to be done there, to translate our "dynamic" representation. To solve
these matters, we consider there are two kinds of `Parameter`s, a
binding (which parses `BParameter`) and a pattern (which parses
`Pattern`). The former is only used by PascaLIGO. A small parser was
made afterwards, and gets used whenever we have a `functionScopedDecl`
call. The grammar for PascaLIGO was changed to more correctly parse
patterns, and tests were changed accordingly.

On top of that, the signature help also changed a bit, since it was not
compatible with our current format anymore. It seems that it expects
only function arguments, but not changes such as tuples, and that was
changed so it would type-check in the first place (since before it
expected `Text`, and now we give it `Parameter`).