[LIGO-271] Handle curried functions
Problem: Seems like we don't handle curried functions correctly in
signature help with CameLIGO. We only see the "first argument".

Solution: Change the grammar to parse as many arguments as it can. Some
operator precedences were wrong so fix that along the way. Create
`_core_expr` as in the reference grammar so the parsing of multiple
arguments work as intended.