[LIGO-263] Add keywords to completion
Problem: We want to make VSCode aware of keywords for better
completions.

Solution: Add support for keywords by refactoring Completion a bit.
Particularly, we want to distinguish ordinary completions from keyword
ones. We create sets containing all keywords and filter only the ones
we're interested in (they are a subsequence of what the user typed).
Change our tests to also include keywords.

On the way, refactor our test driver for Completion a bit so it's not
quadratic anymore.