[LIGO-271] Use a smarter strategy for grouping tuples
Problem: The previous approach to the problem has shown some weaknesses
that need to be addressed. For instance:
* What if we are giving a non-tuple argument to a tuple parameter?
* What if we have mixed tuples with curried functions?

Solution: First of all, we need a way to express parenthesis in
patterns. So we add `IsParen` to our skeleton and do appropriate
changes. Now, in SignatureHelp, we can get our parameters and arguments
and iterate them trying to match whether we have tuple/tuple,
tuple/value, value/tuple or value/value. While we're at it, we extract
the current active index, if it exists. This will give us a list of
tuple arguments, which we can now give to LSP as our parameters.
Meanwhile, our label is still the "old" parameters for a prettier/more
accurate output.