[LIGO-275] No duplicate fetch for changed document
Problem: When we change a document, it ends up calling two handlers: the
one for changed text document, and the one for watched files. This
causes a document to be fetched twice.

Solution: Only fetch documents in watched files if it's a document that
is _not_ open, since open ones will inevitably be called by the other
handler.