[LIGO-275] Use cached version of contracts
Problem: After we edit a document, I imagine the following will happen:
* Text document did change will reparse the contract
* Text document signature help will reparse the document
* Text document code action will reparse the document
* Text document symbol will reparse the document

All of this contributes to the unresponsiveness of the extension.

Solution: I've decided to attribute how much effort each thing should do
to fetch contracts (least, normal, best). In the case of least effort,
we return whatever we have, but begin fetching a new one in background
if we need. I've refactored RIO a bit to make things nicer to use.