[LIGO-279] Fix error nodes in LIGO's for
Problem: When we have a for in LIGO, we're putting an error node, and
causing many tests to fail.

Solution: Turns out our parser is ok, but our recognized isn't. We try
to parse its collection field, but there are no cases for it. We simply
add a new `Collection` with alternatives for map/set/list and recognize
it.