[LIGO-280] Uncomment k.ligo test
Problem: Turns out we accidentally fixed this test with LIGO-271, but we
forgot to uncomment this.

Solution: Uncomment the test. Regarding LIGO-271, seems like it was
fixed with 65beac7e6cfcad2d9b15df9a985d58cd96d38595 where we use
`_param_pattern` instead of directly matching on the wildcard. I assume
it happened because our recognizer has no case for `NameWildcard`, but
since we changed to `wildcard_pattern`, the issue was fixed.