[LIGO-275] Better timestamp in load
Problem: In `load`, we are supposed to give the timestamp of when we
started to load. But instead, we use the timestamp of when we finished
loading. This causes our `ASTMap` to always think our loaded results are
the most up-to-date ones, but that might not be true.

Solution: Create a helper `getTimestamp` function and use it accross the
codebase. Export it. Let `insert` accept the timestamp. Get the time
when we start loading, and use such timestamp to insert in the `ASTMap`.
This wait, we now won't override results if we end up seeing newer ones
while we were loading, in the meantime.