[LIGO-275] Less MVar locking
Problem: `load` "locks" everything with an `MVar` while doing its stuff,
but we only need such locking while generating the inclusions graph.

Solution: Move operations to fetch inclusions to its own function, to
improve readability. We move the locking function to this new function,
which, while it loads dependencies, will remain locked. This means that
now, loading the root contract and adding scopes will no longer need to
be locked, and can be done in parallel for greater performance.