[Chore] Show error message in Error node
Problem: It's difficult to say why a thing failed. All we see most of
the time is `ERROR []` which tells us too little. Would be nice to
improve this to make debugging better.

Solution: Also show the error message, such as `ERROR "fire the
missiles" []` by changing how an error node is pretty printed.