[LIGO-281] Fix type-level strings
Problem: Type-level strings are part of core types in LIGO dialects, but
we don't properly handle them. In CameLIGO, they only appear in tuples.
In PascaLIGO, they only appear in "Michelson types". In ReasonLIGO, they
appear in "Michelson tuples", but they have no fields to be recognized
in the core pattern, where they also appear. Meanwhile, the reference
grammars accept them anywhere.

Solution: Get rid of all this Michelson stuff, since they are not part
of the reference grammar. For CameLIGO and PascaLIGO, add `string_type`
and recognize them. For ReasonLIGO, simply add a field to it and
recognize it. Change the AST skeleton so that we expect an `it` type
variable instead of directly a String for consistency. Uncomment the
remaining test.