[LIGO-243] Use --infer flag in scopes
Problem: There is an `--infer` flag available in the options for
`ligo get-scope`, which provides type information inferred by the
compiler. We currently don't make use of it.

Solution: Call `ligo get-scope` with the `--infer` flag, extract the
resulting type information, and add it to our scopes. Add corresponding
tests.