[LIGO-306] Support type variables in type declarations
Problem: Type variables are local to type declarations, and we still
want some capabilities, such as references and renaming, to work on
them.

Solution: Change the 'ScopedDecl' parser a bit to to include type
declaration specifics in type parameters, as well as 'ScopedDecl'. Now,
we have two kinds of type variables, the ones that are declared and the
ones that are used as ordinary types, so we change 'TypeDeclSpecifics'
to support which kind of type declaration we are talking about, that is,
whether it's a type variable or an ordinary type. Change 'Fallback' so
that we properly parse these parameters and return their scopes. Change
'BTypeDecl' branches so that we introduce type variables in their local
scopes (instead of globally). Type variables in terms (values and
functions) are unhandled for now.