[LIGO-313] Fix type annotations issues in VSCode
Problem: When we have type annotations, they break with a newline.

Solution: Don't use `$` as end regions for type annotations. Further, in
PascaLIGO, match one or more types instead of zero or more otherwise it
will break as well.

This caused some places to stop highlighting properly. The solution is
to further annotate matches with what they may end with. For example, in
CameLIGO:

```
let x : int =
  type foo = int in
  let y : int = 42 in
  y
```

The code above would have improper syntax highlighting with the type
definition. Previously, it would end with the line break, but now it
needs to be adjusted so that an "in" causes it to stop. Other
adjustments similar to this were made both to CameLIGO and PascaLIGO.
ReasonLIGO had no problems with syntax highlighting w.r.t. newlines.