[LIGO-308] Consistently depend on ligo across tests
Problem: Our tests may yield different results depending on whether they
are running on Nix or Stack. This happens because Nix selectively
chooses when to depend on LIGO or not, while Stack always looks in path.
The result: tests that depend on preprocessing are inconsistent.

Solution: First, make `ligoBinaryPath` impure, by always looking for it
in the environment instead of using TH to generate a compile-time value.
The new default is just "ligo" now, which will look in `PATH`. Change
our tests so that they properly set the environment variable
`LIGO_BINARY_PATH` before running the tests, by explicitly creating
`main` functions for each of them, that explicitly say whether they
depend on LIGO or not.