fix version of ocaml-recovery-parser (0.1)
Problem: we want to avoid updating dependency on every new commit in the
ocaml-recovery-parser and keep its asynchronous development.

Solution: pin version of the package and increase it when needed

Note: the version in the name of `pin-depends` can be arbitrary but it would be
better to keep it the same with the tag.