[LIGO-309] Improve syntax highlighting in VSC for PascaLIGO
Problem: When we have anything that is not a pattern after a `const` or
`var`, the syntax highlighting breaks, and they are not even highlighted
when they are in parameters. Also, unrelated, numeric literals are not
highlighted.

Solution: Instead of having something of form `const foo`, change it to
only `const` and include the `foo` (as `indetifiers`) instead. Seems
like TextMate has better recovery with things like this. Do the same for
functions. In addition, function parameters are now highlighted with the
`parameters` rule, but type highlighting had to be changed a bit so it
stops highlighting when it finds a `,`, otherwise multi-param functions
would be incorrectly highlighted. For numeric literals, just include it
on top level patterns.