[LIGO-356] Support disabling features
Problem: We should allow the server to enable or disable specific
features.

Solution: In `package.json`, support a new configuration for disabling
features by its name. In the server, update `Config` to contain such
list of disabled features, and if we find such feature, we answer VSCode
that it was cancelled.