[LIGO-325] add approximate location to recovered token
Problem: now synthesized by error recovery tokens have ghost locations that can
be confusing

Solution: remember the last token that the parser consumed and use its end as
the start of the synthesized token. So it gives a good approximation of the
potential location for missing tokens.

Other improvements: align recovery annotations in the `ParToken.mly`