[LIGO-358] Fix type annotations in lambdas
Problem: We currently can't write type annotations for the return type
of CameLIGO lambdas.

Solution: We try to mimic a bit how the upstream parser is structured.
In some cases, we want to have parenthesis around a type and in some
cases we don't, in order to avoid a LR(1) conflict. `sum_type` and
`variant` were extracted into a non-orthodox "template", which allows us
to choose how to parse the type inside of the variant. Then when we have
a `sum_type`, we choose it accordingly to the type parser we want.
Refactor types according to upstream in order to accomplish this. Remove
the relevant contract from the ignore list. Adjust other tests
accordingly.