[LIGO-323] Disambiguate cons expr from directives
Problem: If we have something something like `const x = a < b` followed
by `#include "foo.ligo"`, our parser trips up and thinks the `#` is the
cons operator, not a directive.

Solution: I'm not sure why, but using Regexes solved the problem. This
is not even the first time switching between Regexes and strings+DSL
helped.