[LIGO-259] Use version of lsp with withProgress fix
Problem: There is a bug in `withProgress` that prevents it from being
used.

Solution: Bump lsp to a newer version that contains the fix. Please see
https://github.com/haskell/lsp/pull/372. Note: Numeric types were
changed in this version, so many changes had to be made to accomodate
it. I decided to change the fields in `Range` to use `Word32`, since
that's what `lsp` expects. Furthermore, `tree-sitter` also uses `Word32`
in some of its types, and so byte-offset was changed as well. I'm not
sure about ligo itself, but I hope `Word32` will suffice there as well.