[LIGO-259] Use withProgress for long running functions
Problem: It would be nice to have a notification indicating what the
extension is currently doing.

Solution: Refactor parsers and scopers to support a callback. In
parsers, this is simpler, just report the current contract number
(zipped with the contract). For scopes, we use an `MVar` to update a
counter due to best practices of `MonadUnliftIO`. We support
cancellation by trying to remember what we've seen so far for parsed
contracts. For scoped contracts, we just cancel everything and load the
default one. I'm not sure whether this is the best approach, however
(compared to just now allowing cancelling, for example). Propagate
changes to tests.