[LIGO-219] Properly clear diagnostics
Problem: The way we were clearing diagnostics didn't work as we expect
it should, as we never specified a source, which caused diagnostics to
just stay there until a new problem appeared most of the times.

Solution: Add the correct source when clearing diagnostics (essentially,
say the source has no diagnostics).