[Chore] Get rid of Product in RioEnv
Problem: When we can write `asks getElem` it's a nice shortcut, but when
we can't, we need to type the entire type signature, which is pretty
annoying. But overall, I don't really like using Product in the place of
an ordinary record, as it makes the implementation more obscure and
probably isn't as efficient as a record, besides that records will have
better error messages when something is wrong.

Solution: Replace it with an ordinary record, propagate changes.