[LIGO-219] Call preprocess on unsaved buffer
Problem: We can't call `ligo print preprocessed` on an unsaved buffer.

Solution: Make a temporary file, dump the contents of the file we want
to preprocess in it. Now we preprocess this file, but now line markers
and errors will be messed up. Replace their file names with the actual
file name we had.

While debugging this, I found that some preprocessing errors will cause
JSON to fail parsing due to the fact they are a string, not an object.
Implement a new case so it can parse either of them.