[LIGO-359] (Actually) release v0.4.6
Problem: Commit 351a250002 bumps the version number of the VSCode
extension from 0.4.5 to 0.4.6, but because of issues with the VSCode extension
marketplace API, we didn't actually release v0.4.6 then.

Solution: Create a little commit that just perturbs the line containing
the version number, while keeping the actual version at 0.4.6. This way,
`git blame` works properly, the commit history properly reflects the
timeline of when 0.4.6 was released.