Fixup a version of the ocaml-recovery-parser
Problem: the last tag (2.0) is placed on too old commit. Also, when the tag has
been bumped in the `ligo.opam`, the version of the package was not updated.
A combination of this two bugs hides it in the CI that looks at version of
the package.

Solution: properly bump the version of the ocaml-recover-parser up to a new tag.

Other improvements:
* add `--` into install_vendors_deps.sh to fix parse error during local build
  via Makefile.
* new version of the ocaml-recovery-parser contains a fix of the license-field