[LIGO-379] Remove global references from the parser frontend
Problem: in the current version of `ocaml-recovery-parser` there is only one way
to pass approximate location of synthesized tokens into recovery grammar
annotation — global reference.

Solution: extend API of the `ocaml-recovery-parser` by meta variable `$loc` (see
MR#10) and use it instead of global references to location.
Also since we start getting a position from `Merlin_recovery` we have to
properly convert `Pos.t` to `Lexing.position` and vise versa, but it is lossy
and it's possible to keep correct whether only byte or only code point position.
So handling the `--bytes` lexer option in the ParserLib helps to choose which
one to keep.

Other improvements:
* fix a bug in the https://gitlab.com/ligolang/ligo/-/merge_requests/1469
that causes an exception during the error recovery next to ghost token (in the
reasonligo)
* fix ignoring `--bytes` in the LexerLib