[LIGO-382] Use UnliftIO.Exception instead of Control.Exception.Safe
Problem: We currently use three libraries that support exceptions:
`exceptions`, `safe-exceptions` and `unliftio`. Let's only use one of
them: `unliftio`. Note: FPComplete (fpco) are the authors of both
`safe-exceptions` and `unliftio` and recommend using `unliftio`. This
information is mentioned in the Hackage packages of both projects, as
well as their blog posts about these packages, and hence I preferred
this package, besides that we've already been depending on it for a
while.

Solution: Remove package dependencies. Propagate changes trivially,
using the alternatives in `UnliftIO.Exception`.