[LIGO-382] Don't try to recover from any exception
Problem: We try to recover from any synchronous exception that may
occur, including ones that shouldn't, such as ones caused by programmer
errors or unrecoverable ones, like `HardwareFault` or `NonTermination`.

Solution: Let the extension die if we have an uncaught exception. This
is a bug, and should be reported to us. It's also better than leaving
the extension in an inconsistent state. Note that the LSP protocol says
the client should deal with server crashes, by attempting to restart it
a few times, but I can't find where I've read it anymore. VS Code, for
example, tries to restart the server 5 times if it died within 3
minutes, otherwise leaves it dead.