[LIGO-176] Delete Debouncer
Problem: We haven't used the debouncer for a few months now, and our
previous problems with MVar haven't appeared again. Furthermore, trying
to enable the debouncer once again causes many tests to start failing,
as there is something odd going on with it. Our ASTMap also handles that
everyone gets "the result of the winner", which makes this module pretty
much irrelevant. See the issue page for more information, where a
discussion was created about it.

Solution: Delete it.