[LIGO-382] Improve Failure exception
Problem: Failure uses a `String` field with the error, and it should be
more structured. Also one of our tests try to catch this exception and
use it as "expected", but this exception is a programmer error and
should not be expected. Same for ScopeError, which isn't even thrown in
the first place.

Solution: Rename `Failure` to a more descriptive name:
`UnrecognizedFieldException`. Instead of saving the error message, save
the name of the field that failed. For our test, just don't `try`
catching it in the first place, and don't catch `ScopeError` either. Add
some more information to `UnrecognizedFieldException` so it's easier to
debug. Refactor duplicate code in `boilerplate`.