[LIGO-382] Don't recover from TreeDoesNotContainName
Problem: `ScopeError` is a programmer error, not an user error, and we
shouldn't recover from it.

Solution: First, move `ScopeM` and `ScopeError` to `AST.Scope.Fallback`,
as this is the only module that uses these types. Rename `ScopeError`
and `TreeDoesNotContainName` both to `TreeDoesNotContainNameException`.
Throw it from `Fallback` scopes as needed.