[LIGO-212] Don't overlap Show with Pretty
Problem: We have a pretty annoying orphan instance where `Show` is
automatically derived for things with a `Pretty` instance. Normally,
it's a pain when we want to show something for debugging. I believe that
`Show` should be programmer-friendly, while `Pretty` should be
user-friendly, but this separation doesn't exist here.

Solution: Delete the overlapping instance. After this,
`ContractNotFoundException` broke, and I had to make some changes. I
changed it so that what was the old `Pretty` instance becomes the
`displayException` method of `Exception`, and `Show` is automatically
derived.