[LIGO-212] Make ParseTree Show-friendly
Problem: ParseTree uses a Pretty instance which I've never ever got to
work in the first place, so why not use just a plain Show?

Solution: Get rid of Pretty instance, use an equivalent Show instance. I
opt to not show the source, since it's too verbose for a
programmer-friendly output (and it's already quite verbose as it is). I
got rid of the Product and used a tuple while I were at it, for
simplicity.