[LIGO-212] Some more Pretty and Show instances
Problem: Sometimes GHCi complains when trying to Show some things, so
let's add some more instances.

Solution: Change Duplo to add a ShowPP, which allows deriving Pretty via
Show. I decided to keep some Pretty instances as the default for Show
for now, since they are shorter and easier to read.

I took the liberty to slightly change the Pretty instances for Range and
DeclRef in a way I think that is a bit easier to read.