[LIGO-212] Prefer displayException to pp in Exceptions
Problem: There is no reason for us to derive Pretty for Exceptions, and
we should prefer displayException for this purpose. The Show instance
should be helpful for programmers, while displayException should be
helpful for programmers.

Solution: Don't derive Pretty for LIGO exceptions, but rather use
displayException. Reserve the Show instance to be auto-generated, to be
easier to parse by a programmer.