[LIGO-226] Improve completion kind
Problem: Up until now, we only returned `CiFunction` as the completion
kind for all identifiers (and `CiKeyword` for keywords). It's ideal that
we actually use the correct completion kind.

Solution: Figure out the kind from the scope, based on the type that was
provided from the `TypeDeclSpecifics`. In case of a function, we check
if there are any parameters first, and return `CiFunction` in this
situation. For fields, we return `CiField`.