[LIGO-226] Provide a completion kind for constructors
Problem: We also want to be able to correctly provide a kind for
constructors. Due to naming conventions, if it starts with an upper
case letter, it's likely a constructor, so we can easily figure it out.
Note, VS Code provides the same icons for functions and constructors as
of the time of this writing.