[LIGO-419] Fix format reverting code
Problem: When we call format, we do so in the file which is in disk to
call LIGO, but it reads an old version and ends up reverting the
progress of the user.

Solution: Dump the contents of the unsaved buffer to a temporary file
before calling LIGO in it.