[LIGO-425] Fix broken comment insertion
Problem: In VS Code, when the user tries to add a comment using a
shortcut, such as "Ctrl+C Ctrl+A" or "Ctrl+Shift+A", it prints a regex
instead.

Solution: Have separate sections for comment insertion and comment
highlighting. Keep comments as it is, but use ordinary (non-regex)
strings for insertion.