[LIGO-409] refactor recovery tests driver, create special folders
Problem: currently we do not check validity of the recovery output. Also, not
all tests would pass this check because of errors before recovery.

Solution: extract tests on which lexer or preparser fails into own folder and
so exclude it from CI. Add the check that output of the recoverable parser have
no syntax error.