[LIGO-430] Don't discard contracts to load them again
Problem: We load a bunch of `Source`s, discard them, and later try to
load them again.

Solution: Keep the `Source`s, just parse them and insert preprocessor
ranges as well as errors.

Although I'm not quite sure if this is the best solution, as we avoid
some of our mechanisms, like `preload`.