[LIGO-436] Fix decoding of module alias' right hand side
Problem: There are two problems: first that we say the module alias may
be 0 or more names, when it's actually 1 or more. Second that we try to
decode it as a single field.

Solution: Use `sepBy1` instead of `sepBy` in grammars to decode 1+. Make
the AST expect a list of names instead of only one, and use `fields`
instead of `field`.