[LIGO-436] Attempt to fix invalid decoding of nodes
Problem: Seems like we have an interesting bug: if we have some error in
a node, and this node's name contains a string field with the same name
as the node, we somehow decode it incorrectly (as the keyword) without
its children in the fields. This appeared in some recovery tests, and
seems like only renaming the node was enough to fix this. Hopefully this
will also resolve LIGO-380, since it seems like it was caused exactly by
`if` from ReasonLIGO dialect.

Solution: Rename some node names to not overlap with its keywords. Note
that there may be more missing cases, but I haven't been able to find
more cases, so it seems fine.