[LIGO-412] Add FromCompiler tests where possible
Problem: We want to check that FromCompiler tests work as expected.

Solution: Add FromCompiler to integration tests, in addition to Standard
ones.

n.b.: A lot of tests are actually failing, they seem to be caused by
LIGO-93 (in rename and find tests mostly), LIGO-208 (in various tests),
and LIGO-446 (in a few PascaLIGO tests).