[LIGO-257] Add an entrypoint entry button
Problem: We want to have an ability to specify
entrypoint, which we are going to compile.

Solution: Add a textbox to write entrypoint name to call
when compiling contract

Note: This commit also fixes a bug with hiding
button in non-ligo files. When you switch active text
window for example for terminal, compilation command
will be executed with bad path. In order to prevent
it, extension will first check, that given file is
actually ligo file, not some vscode window.