[LIGO-446] Fix malformed upstream contracts
Problem: Three contracts in upstream can't be parsed by LIGO or by us,
but they seem like they should be.

Solution: Fix them so they are not malformed anymore. `match.ligo` is
meant to be broken, and probably the ones in `negative` as well,
although I'm not sure about the negative ones.