[LIGO-430] Cache inclusions graph (only file names)
Problem: We forget how to parse the directory and need to index it again
when we open a new contract.

Solution: Cache at least what were the file names. Instead of indexing
the entire directory again, lookup in our graph. I moved
`loadContractsWithDependencies` to RIO since it became too specific to
be in `AST.Parser`.