[LIGO-467] Break RIO into smaller modules
Problem: RIO has grown to be a quite large module and it's difficult to
navigate around it.

Solution: Split it into a file for
loading/deleting/invalidating/fetching documents, one for diagnostics,
one for LSP stuff, one for types, and one that is the "main" module.

Note: Configuration stuff are expected to be removed in LIGO-411, so I
haven't touched them a lot, although I moved `getCustomConfig` to the
types module to avoid problems with cyclic dependencies.