[LIGO-474] Use TreeDoesNotContainName as a Msg
Problem: We currently use TreeDoesNotContainName as an exception. Our
original plans for throwing exceptions was roughly that we want
programmer errors to be thrown as exceptions, and user errors to become
a diagnostic. Although TreeDoesNotContainName can be caused by
programmer errors, it can also appear as legitimate user errors, and so
we shouldn't throw it.

Solution: Accumulate such error in the Writer, add it to the list of
errors at the end.