[LIGO-449] Decode new get-scope output
Problem: The output for get-scope now returns a list of errors and
warnings, in addition to the definitions.

Solution: Change JSON to contain both errors and warnings. Create new
LigoMessage datatype for the case when we get no definitions (LIGO
failed). Properly decode them and adapt datatypes and code to propagate
changes.