[LIGO-397] Create local temporary directory
Problem: We want a temporary directory to be created locally and avoid
`/tmp`, as Docker won't be able to read files from `/tmp`.

Solution: Pass around a new `.ligo-work` directory, which will serve as
the temporary directory. It tries to be created in the same place as
where `.ligoproject` might be found, but if it isn't there, then just
put it in the same directory as the file we're currently looking. Also,
add a mechanism to ignore paths, so we don't accidentally index anything
in the ignored directory. This also has the advantage that one of our
test suites was simplified, and it paves the way for LIGO-491 which will
also require this ignoring mechanism.