[LIGO-397] Spawn LIGO within the index directory
Problem: We should spawn LIGO from within the index directory, so that
it doesn't complain about being in root or something.

Solution: Pass index directory as cwd. For the Visual Studio Code
extension, implement a request that returns the index directory (if any)
and use it to figure out the directory.