[LIGO-303] Start using stylish-haskell
Problem: the formatting of imports is currently quite custom, it would
be nice to have `stylish-haskell` config and auto-format according to
it.

Solution: add `.stylish-haskell.yaml` that was copied from LSP project
with small changes to follow the current style is close as possible.

Still, there is no way to leave the enclosing `)` of import lists at the
new line with stylish-haskell, so I changed this point of the current
style.