[LIGO-488] Add syntax highlighting for ligo
Problem: we need an ability to highlight ligo syntax

Solution: mligo and ligo already have old syntax highlighting
rules in monaco. And religo with jsligo may work with js syntax.

Limitation: this syntax highlighting is old and we need to
improve it in the future. Also we disabled decorations for
js as religo with jsligo different from real js code.