[LIGO-411] Add logic to process config updates
Problem: We need to update the configuration on initialization, as well
as whenever we receive a workspace/didChangeConfiguration notification.

Solution: To set configuration at initialization, we do it while setting
up the RIO monad. To handle configuration updates, we implement a
handler for the workspace/didChangeConfiguration request. Note that
implementing onConfigurationChange is neither useful nor necessary: in
VSCode, the client sends configuration change notifications with Null
bodies, and the server is expected to fetch the full configuration.
Since onConfigurationChange is a pure function, it cannot perform this
fetch, so it's useless when the client is VSCode.