[LIGO-411] Test configuration management
Problem: We need to test whether we are updating the configuration
properly.

Solution: There is already a test `unit_can't_hover_when_disabled` that
tests that configuration is initialized properly, by disabling the hover
capability in the initial configuration and ensuring that the server
respects this setting. We add a similar test
`unit_can't_hover_after_disabling` to ensure that the server respects
this setting if it comes as a configuration update as well.