[LIGO-397] Add a way to filter contracts
Problem: We will need to filter contracts based on whether they are in a
temporary directory or not.

Solution: Add a filtering function while scanning contracts, which
doesn't even visit a directory if its predicate failed. Furthermore,
extract some functions from the parser that are only used in tests to
the test suite.