[LIGO-544] Fix bug with deployment
Problem: deploy commands stopped working.It seems
that it is happening, because keys, generated for
tezos origination. After ithacanet was. launched,
support for hangzhounet was canceled, but it is still
hardcoded into the code for operations

Solution: Add a choice of only two protocols for
deployment operations - currently active protocol and
maainnet. Currently active protocol can be specified
in settings