[LIGO-491] Ignore files from the indexing mechanism
Solution: We'd like to amend `.ligoproject` to avoid indexing certain
files and directories.

Solution: Turn `.ligoproject` into a JSON file which allows for an
optional field to ignore dirs. While indexing, do not look for files
that are on the ignore list. Also, do not index if we're opening an
ignored file. Add some tests.