[LIGO-460] Fix vscode warning
Problem: Extension sometimes shows warnings, that
root directory is not specified (or to be precise that it is Nothing).
This is happening because indexing always try to get
directory, while in fact the. choice. is still pending.

Solution: Shut this warning in case. index directory
choice is pending, or. if we choose to do not index.