Use removal of source mappers
Problem: source mappers turned out to be not strictly necessary, and
recent `morley-debugger` does not have them already.

Solution: re-write the logic to avoid them.

We still need to get all the notable source locations for the sake of
breakpoints (debugger needs to know where it is possible to set a
breakpoint). But the indirection related to carrying `InstrNo`s and then
accessing source locations through them is now gone.