Improve one error message
Problem: the error about too many debug entries is very verbose, but the
list it returns is not informative.

Usually the reason of this error is that we had to skip some entries
in the middle, and showing few last ones does not help at all.

Solution: instead of the list of remaining debug entries, include number
of these entries into the error.