[LIGO-553] Better Michelson iterop
Problem: The way we currently handle Michelson code injections is not
the best. We don't use `Verbatim` expressions and patterns (called
`MichelsonCode` in our code for some reason) in all places where they
may appear, and we expect a list of arguments in the injections while
the Menhir grammars don't have such thing.

Solution: Rename `michelson_code` to `Verbatim` in CameLIGO and
ReasonLIGO and accept them in patterns and expressions. In PascaLIGO
their usage is already correct, after the grammar rewrite, and in
JsLIGO, only renaming was needed, as this parser doesn't allow them to
appear in arbitrary expressions and patterns.

Delete the arguments, as they aren't really valid LIGO code. Also delete
them from `Michelson` in the Haskell portion of the code.

Rename `MichelsonCode` to `Verbatim` in the Haskell portion of the code.